Combine ActiveStarship and ActiveControlledStarship
